trust-backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
openpyxl==3.1.2
xlwt==1.3.0
